feat(product): 添加SPU商品保存功能并启用Feign客户端
- 在GulimallProductApplication中启用Feign客户端支持
- 新增ProductAttrValueService接口及实现类用于保存商品属性值
- 新增SpuImagesService接口及实现类用于保存商品图片
- 新增SpuInfoDescService接口及实现类用于保存商品描述
- 实现SpuInfoServiceImpl中的saveSpu方法，完成SPU信息、描述、图片和规格参数的保存
- 创建SpuBoundTo传输对象用于积分设置
- 优化DTO中的attrId数据类型从int改为Long以支持更大范围的ID值

+++ gulimall-Parent/services/gulimall-product/src/main/java/com/xjz/gulimall/product/dto/SpuSaveDto.java
@@ -62,7 +62,7 @@ public class SpuSaveDto {
     }
     @Data
     public static class BaseAttrs {
-        public int attrId;
+        public Long attrId;
         public String attrValues;
         public int showDesc;
     }
